+++ packages/in_app_purchase/in_app_purchase_storekit/example/ios/RunnerTests/Stubs.m
@@ -1,1 +1,0 @@
-../../shared/RunnerTests/Stubs.m