+++ packages/in_app_purchase/in_app_purchase_storekit/example/macos/RunnerTests/RunnerTests-Bridging-Header.h
@@ -2,4 +2,5 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import "Stubs.h"
+#import <FlutterMacOS/FlutterMacOS.h>
+@import in_app_purchase_storekit_objc;